requirements/requirements.txt
requests>=2.18.4
grpcio==1.19.0
google-cloud-storage
google-cloud-kms
flask
backtrader
iexfinance>=0.4.3
